requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
requests==2.32.3
Werkzeug==3.0.4
python-dotenv==1.0.1
Pillow==12.1.1
pytest==8.3.3
pytest-flask==1.3.0
